♻️ Extract Checkout endpoint host into CheckoutEndpointHost
